Include Emoji Support
This commit adds emoji support. Now, the TorDice class will grab the custom emojis out of the Guild object and then look for specific emojis (of the form d6X and d12X and d61-3weary). It then creates the full string necessary to print this emoji and stores it in  a container class. The compile string then calls helper functions for getting the dice strings that will look up the relevant string in the container class.