Fixed Issue Adversary Dice
Added the ability to have adversary dice provide a different emote. These emotes are keyed as "aX", where "X" is the regular die emote name (such as "d121", d63weary", etc.). If no adversary die emote exists for a given die result, then the regular emote is used.

Additionally, changed the background implementation of the emote interface to use a hashtable. This hashtable is shared across all instances of the EmtoeStrings class; saving memory and only needing to be updated once per server. 

The hashtable is keyed on the String ID of the server, and provides another hashtable. This hashtable is keyed on the emote name (such as "d121", and provides the full emote string for the emote on the server of the containing entry. For a given server, it's hashtable entry is only updated on the first roll on that server after the bot is started, unless the new command "reloadEmotes" is issued, which will force a reload of the hashtable. This is to support changing the emojis for the dice without having to restart the bot.